+++ CHANGELOG.md
@@ -75,6 +75,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   conflicts to be materialized and parsed correctly in files which already
   contain lines that look like conflict markers.
 
+* New `$marker_length` variable to allow merge tools to support longer conflict
+  markers (equivalent to "%L" for Git merge drivers).
+
 ### Fixed bugs
 
 * The `$NO_COLOR` environment variable must now be non-empty to be respected.